Moves some TipKit code to BSK to share with macOS
