+++ DuckDuckGo/NetworkProtectionStatusViewModel.swift
@@ -25,6 +25,7 @@ import BrowserServicesKit
 import Core
 import Subscription
 import TipKit
+import TipKitUtils
 
 struct NetworkProtectionLocationStatusModel {
     enum LocationIcon {
