--- DuckDuckGo/TipKit/TipKitController.swift
@@ -1,110 +1,0 @@
-//
-//  TipKitController.swift
-//  DuckDuckGo
-//
-//  Copyright © 2024 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import Core
-import Foundation
-import os.log
-import TipKit
-
-protocol TipKitControlling {
-    @available(iOS 17.0, *)
-    func configureTipKit()
-
-    @available(iOS 17.0, *)
-    func resetTipKitOnNextAppLaunch()
-}
-
-typealias TipKitAppEventHandler = TipKitController
-
-final class TipKitController {
-
-    private let logger: Logger
-    private let userDefaults: UserDefaults
-
-    private var resetTipKitOnNextLaunch: Bool {
-        get {
-            userDefaults.bool(forKey: "resetTipKitOnNextLaunch")
-        }
-
-        set {
-            userDefaults.set(newValue, forKey: "resetTipKitOnNextLaunch")
-        }
-    }
-
-    init(logger: Logger = .tipKit,
-         userDefaults: UserDefaults = .standard) {
-
-        self.logger = logger
-        self.userDefaults = userDefaults
-    }
-
-    @available(iOS 17.0, *)
-    func configureTipKit() {
-        do {
-            if resetTipKitOnNextLaunch {
-                resetTipKit()
-                resetTipKitOnNextLaunch = false
-            }
-
-            try Tips.configure([
-                .displayFrequency(.immediate),
-                .datastoreLocation(.applicationDefault)
-            ])
-
-            logger.debug("TipKit initialized")
-        } catch {
-            logger.error("Failed to initialize TipKit: \(error)")
-        }
-    }
-
-    @available(iOS 17.0, *)
-    private func resetTipKit() {
-        do {
-            try Tips.resetDatastore()
-
-            logger.debug("TipKit reset")
-        } catch {
-            logger.debug("Failed to reset TipKit: \(error)")
-        }
-    }
-
-    /// Resets TipKit
-    ///
-    /// One thing that's not documented as of 2024-10-09 is that resetting TipKit must happen before it's configured.
-    /// When trying to reset it after it's configured we get `TipKit.TipKitError(value: TipKit.TipKitError.Value.tipsDatastoreAlreadyConfigured)`.
-    /// In order to make things work for us we set a user defaults value that ensures TipKit will be reset on next
-    /// app launch instead of directly trying to reset it here.
-    ///
-    @available(iOS 17.0, *)
-    func resetTipKitOnNextAppLaunch() {
-        resetTipKitOnNextLaunch = true
-        logger.debug("TipKit will reset on next app launch")
-    }
-}
-
-extension TipKitController: TipKitAppEventHandling {
-
-    func appDidFinishLaunching() {
-        if #available(iOS 17.0, *) {
-            configureTipKit()
-        } else {
-            logger.log("TipKit initialization skipped: iOS 17.0 or later is required.")
-        }
-    }
-}
